HARMONY-1378: Fix string interpolation and table/example labeling
